requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
idna==2.10
peewee==3.14.4
PyMySQL==1.0.2
python-dotenv==0.17.1
requests==2.25.1
sentry-sdk==1.1.0
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.5
user-agent==0.1.9
